requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-snowflake
apache-airflow-providers-common-sql
pendulum
requests>=2.31
pandas<2.2,>=2.1